+++ Example/ExampleApp/CompanyViewController.swift
@@ -12,7 +12,7 @@
 //
 //
 //  License
-//  Copyright © 2015 Jesse Squires
+//  Copyright © 2015-present Jesse Squires
 //  Released under an MIT license: https://opensource.org/licenses/MIT
 //
 
@@ -36,9 +36,9 @@ final class CompanyViewController: UITableViewController, NSFetchedResultsContro
         showSpinner()
 
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
+        let factory = CoreDataStackProvider(model: model)
 
-        factory.createStack { (result: StackResult) -> Void in
+        factory.createStack { result in
             switch result {
             case .success(let s):
                 self.stack = s
@@ -103,7 +103,7 @@ final class CompanyViewController: UITableViewController, NSFetchedResultsContro
     func didTapAddButton(_ sender: UIBarButtonItem) {
         stack.mainContext.performAndWait {
             _ = Company.newCompany(self.stack.mainContext)
-            self.stack.mainContext.save(wait: true)
+            self.stack.mainContext.saveSync()
         }
     }
 
@@ -116,7 +116,7 @@ final class CompanyViewController: UITableViewController, NSFetchedResultsContro
                 for each in objects {
                     backgroundChildContext.delete(each)
                 }
-                backgroundChildContext.save(wait: true)
+                backgroundChildContext.saveSync()
             } catch {
                 print("Error deleting objects: \(error)")
             }
@@ -162,7 +162,7 @@ final class CompanyViewController: UITableViewController, NSFetchedResultsContro
             stack.mainContext.performAndWait {
                 self.stack.mainContext.delete(obj)
             }
-            stack.mainContext.save(wait: true)
+            stack.mainContext.saveSync()
         }
     }
 
@@ -206,6 +206,8 @@ final class CompanyViewController: UITableViewController, NSFetchedResultsContro
         case .move:
             tableView.deleteRows(at: [indexPath!], with: .fade)
             tableView.insertRows(at: [newIndexPath!], with: .fade)
+        @unknown default:
+            fatalError("Unknown change type \(type)")
         }
     }