+++ Example/ExampleApp/EmployeeViewController.swift
@@ -12,7 +12,7 @@
 //
 //
 //  License
-//  Copyright © 2015 Jesse Squires
+//  Copyright © 2015-present Jesse Squires
 //  Released under an MIT license: https://opensource.org/licenses/MIT
 //
 
@@ -42,7 +42,7 @@ final class EmployeeViewController: UITableViewController, NSFetchedResultsContr
     @IBAction func didTapAddButton(_ sender: UIBarButtonItem) {
         stack.mainContext.performAndWait {
             _ = Employee.newEmployee(self.stack.mainContext, company: self.company)
-            self.stack.mainContext.save(wait: true)
+            self.stack.mainContext.saveSync()
         }
     }
 
@@ -54,7 +54,7 @@ final class EmployeeViewController: UITableViewController, NSFetchedResultsContr
                 for each in objects {
                     backgroundChildContext.delete(each)
                 }
-                backgroundChildContext.save(wait: true)
+                backgroundChildContext.saveSync()
             } catch {
                 print("Error deleting objects: \(error)")
             }
@@ -125,7 +125,7 @@ final class EmployeeViewController: UITableViewController, NSFetchedResultsContr
             stack.mainContext.performAndWait {
                 self.stack.mainContext.delete(obj)
             }
-            stack.mainContext.save(wait: true)
+            stack.mainContext.saveSync()
         }
     }
 
@@ -169,6 +169,8 @@ final class EmployeeViewController: UITableViewController, NSFetchedResultsContr
         case .move:
             tableView.deleteRows(at: [indexPath!], with: .fade)
             tableView.insertRows(at: [newIndexPath!], with: .fade)
+        @unknown default:
+            fatalError("Unknown change type \(type)")
         }
     }