+++ Tests/ResetStackTests.swift
@@ -12,7 +12,7 @@
 //
 //
 //  License
-//  Copyright © 2015 Jesse Squires
+//  Copyright © 2015-present Jesse Squires
 //  Released under an MIT license: https://opensource.org/licenses/MIT
 //
 
@@ -20,6 +20,8 @@ import CoreData
 import ExampleModel
 @testable import JSQCoreDataKit
 import XCTest
+
+// swiftlint:disable force_try
 
 final class ResetStackTests: TestCase {
 
@@ -33,7 +35,7 @@ final class ResetStackTests: TestCase {
         let expectation = self.expectation(description: #function)
 
         // WHEN: we attempt to reset the stack
-        inMemoryStack.reset { (result: StackResult) in
+        inMemoryStack.reset { result in
             if case .failure(let e) = result {
                 XCTFail("Error while resetting the stack: \(e)")
             }
@@ -59,7 +61,7 @@ final class ResetStackTests: TestCase {
         let expectation = self.expectation(description: #function)
 
         // WHEN: we attempt to reset the stack
-        inMemoryStack.reset { (result: StackResult) in
+        inMemoryStack.reset { result in
             if case .failure(let e) = result {
                 XCTFail("Error while resetting the stack: \(e)")
             }
@@ -78,14 +80,14 @@ final class ResetStackTests: TestCase {
     func test_ThatPersistentStore_WithChanges_DoesNotHaveObjects_AfterReset() {
         // GIVEN: a stack and persistent store with data
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
-        let stack = factory.createStack().stack()!
+        let factory = CoreDataStackProvider(model: model)
+        let stack = try! factory.createStack().get()
         let context = stack.mainContext
 
         context.performAndWait {
             self.generateCompaniesInContext(context, count: 3)
         }
-        context.save(wait: true)
+        context.saveSync()
 
         let request = Company.fetchRequest
         let objectsBefore = try? context.count(for: request)
@@ -94,7 +96,7 @@ final class ResetStackTests: TestCase {
         let expectation = self.expectation(description: #function)
 
         // WHEN: we attempt to reset the stack
-        stack.reset { (result: StackResult) in
+        stack.reset { result in
             if case .failure(let e) = result {
                 XCTFail("Error while resetting the stack: \(e)")
             }
@@ -110,3 +112,5 @@ final class ResetStackTests: TestCase {
         XCTAssertEqual(objectsAfter, 0)
     }
 }
+
+// swiftlint:enable force_try