+++ Tests/StackTests.swift
@@ -12,7 +12,7 @@
 //
 //
 //  License
-//  Copyright © 2015 Jesse Squires
+//  Copyright © 2015-present Jesse Squires
 //  Released under an MIT license: https://opensource.org/licenses/MIT
 //
 
@@ -20,6 +20,8 @@ import CoreData
 import ExampleModel
 @testable import JSQCoreDataKit
 import XCTest
+
+// swiftlint:disable force_try
 
 final class StackTests: XCTestCase {
 
@@ -34,9 +36,9 @@ final class StackTests: XCTestCase {
         let sqliteModel = CoreDataModel(name: modelName, bundle: modelBundle)
 
         // WHEN: we create a stack
-        let factory = CoreDataStackFactory(model: sqliteModel)
+        let factory = CoreDataStackProvider(model: sqliteModel)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
 
         // THEN: it is setup as expected
         XCTAssertTrue(FileManager.default.fileExists(atPath: sqliteModel.storeURL!.path), "Model store should exist on disk")
@@ -49,9 +51,9 @@ final class StackTests: XCTestCase {
         let binaryModel = CoreDataModel(name: modelName, bundle: modelBundle, storeType: .binary(URL(fileURLWithPath: NSTemporaryDirectory())))
 
         // WHEN: we create a stack
-        let factory = CoreDataStackFactory(model: binaryModel)
+        let factory = CoreDataStackProvider(model: binaryModel)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
 
         // THEN: it is setup as expected
         XCTAssertTrue(FileManager.default.fileExists(atPath: binaryModel.storeURL!.path), "Model store should exist on disk")
@@ -64,9 +66,9 @@ final class StackTests: XCTestCase {
         let inMemoryModel = CoreDataModel(name: modelName, bundle: modelBundle, storeType: .inMemory)
 
         // WHEN: we create a stack
-        let factory = CoreDataStackFactory(model: inMemoryModel, options: nil)
+        let factory = CoreDataStackProvider(model: inMemoryModel, options: nil)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
 
         // THEN: it is setup as expected
         XCTAssertNil(inMemoryModel.storeURL, "Model store should not exist on disk")
@@ -77,9 +79,9 @@ final class StackTests: XCTestCase {
     func test_ThatChildContext_IsCreatedSuccessfully_WithDefaultParameters() {
         // GIVEN: a model and stack
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
+        let factory = CoreDataStackProvider(model: model)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
 
         // WHEN: we create a child context
         let childContext = stack.childContext()
@@ -94,9 +96,9 @@ final class StackTests: XCTestCase {
     func test_ThatChildContext_IsCreatedSuccessfully_WithCustomParameters() {
         // GIVEN: a model and stack
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
+        let factory = CoreDataStackProvider(model: model)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
 
         // WHEN: we create a child context
         let childContext = stack.childContext(concurrencyType: .privateQueueConcurrencyType, mergePolicyType: .errorMergePolicyType)
@@ -110,9 +112,11 @@ final class StackTests: XCTestCase {
 
     func test_Stack_Description() {
         let model = CoreDataModel(name: modelName, bundle: modelBundle)
-        let factory = CoreDataStackFactory(model: model)
+        let factory = CoreDataStackProvider(model: model)
         let result = factory.createStack()
-        let stack = result.stack()!
+        let stack = try! result.get()
         print(stack)
     }
 }
+
+// swiftlint:enable force_try